converted u64 to u128 to allow for bigger numbers

+++ src/markets.rs
@@ -13,7 +13,7 @@ struct Markets {
 	creator: String,
 	active_markets: BTreeMap<u64, Market>,
 	nonce: u64,
-	fdai_balances: HashMap<String, u64>, // Denominated in 1e18
+	fdai_balances: HashMap<String, u128>, // Denominated in 1e18
 	fdai_circulation: u128,
 	fdai_in_protocol: u128,
 	fdai_outside_escrow: u128,
@@ -23,13 +23,13 @@ struct Markets {
 #[near_bindgen]
 impl Markets {
 
-	fn dai_token(&self) -> u64 {
-		let base: u64 = 10;
+	fn dai_token(&self) -> u128 {
+		let base: u128 = 10;
 		return base.pow(17);
 	}
 
 	// This is a demo method, it mints a currency to interact with markets until we have NDAI
-	pub fn add_to_creators_funds(&mut self, amount: u64) {
+	pub fn add_to_creators_funds(&mut self, amount: u128) {
 		let from = env::predecessor_account_id();
 		assert_eq!(from, self.creator);
 
@@ -56,7 +56,7 @@ impl Markets {
 		return true;
 	}
 
-	pub fn get_fdai_balance(&self, from: String) -> u64 {
+	pub fn get_fdai_balance(&self, from: String) -> u128 {
 		return *self.fdai_balances.get(&from).unwrap();
 	}
 
@@ -83,7 +83,7 @@ impl Markets {
 		}
 	}
 
-	pub fn place_order(&mut self, market_id: u64, outcome: u64, spend: u64, price_per_share: u64) {
+	pub fn place_order(&mut self, market_id: u64, outcome: u64, spend: u128, price_per_share: u128) {
 		let from = env::predecessor_account_id();
 		let balance = self.fdai_balances.get(&from).unwrap();
 		assert!(balance >= &spend);
@@ -96,7 +96,7 @@ impl Markets {
 		self.subtract_balance(rounded_spend);
 	}
 
-	pub fn cancel_order(&mut self, market_id: u64, outcome: u64, order_id: u64) {
+	pub fn cancel_order(&mut self, market_id: u64, outcome: u64, order_id: u128) {
 		let from = env::predecessor_account_id();
 		let market = self.active_markets.get_mut(&market_id).unwrap();
 		assert_eq!(market.resoluted, false);
@@ -113,7 +113,7 @@ impl Markets {
 		market.resolute(from, winning_outcome);
 	}
 
-	fn subtract_balance(&mut self, amount: u64) {
+	fn subtract_balance(&mut self, amount: u128) {
 		let from = env::predecessor_account_id();
 		let balance = self.fdai_balances.get(&from).unwrap();
 		let new_balance = *balance - amount;
@@ -124,7 +124,7 @@ impl Markets {
 		self.fdai_in_protocol= self.fdai_outside_escrow + amount as u128;
 	}
 
-	fn add_balance(&mut self, amount: u64) {
+	fn add_balance(&mut self, amount: u128) {
 		let from = env::predecessor_account_id();
 		let balance = self.fdai_balances.get(&from).unwrap();
 		let new_balance = *balance + amount;
@@ -135,19 +135,19 @@ impl Markets {
 		self.fdai_in_protocol= self.fdai_outside_escrow - amount as u128;
 	}
 
-	pub fn get_open_orders(&self, market_id: u64, outcome: u64, from: String) -> &BTreeMap<u64, Order> {
+	pub fn get_open_orders(&self, market_id: u64, outcome: u64, from: String) -> &BTreeMap<u128, Order> {
 		let market = self.active_markets.get(&market_id).unwrap();
 		let orderbook = market.orderbooks.get(&outcome).unwrap();
 		return &orderbook.open_orders;
 	}
 	
-	pub fn get_filled_orders(&self, market_id: u64, outcome: u64, from: String) -> &BTreeMap<u64, Order> {
+	pub fn get_filled_orders(&self, market_id: u64, outcome: u64, from: String) -> &BTreeMap<u128, Order> {
 		let market = self.active_markets.get(&market_id).unwrap();
 		let orderbook = market.orderbooks.get(&outcome).unwrap();
 		return &orderbook.filled_orders;
 	}
 
-	pub fn get_claimable(&self, market_id: u64, from: String) -> u64 {
+	pub fn get_claimable(&self, market_id: u64, from: String) -> u128 {
 		return self.active_markets.get(&market_id).unwrap().get_claimable(from);	
 	}
 
@@ -171,12 +171,12 @@ impl Markets {
 		return market.unwrap();
 	}
 
-	pub fn get_market_order(&self, market_id: u64, outcome: u64)  -> Option<u64> {
+	pub fn get_market_order(&self, market_id: u64, outcome: u64)  -> Option<u128> {
 		let market = self.active_markets.get(&market_id).unwrap();
 		return market.orderbooks[&outcome].market_order;
 	}
 
-	pub fn get_market_price(&self, market_id: u64, outcome: u64) -> u64 {
+	pub fn get_market_price(&self, market_id: u64, outcome: u64) -> u128 {
 		let market = self.active_markets.get(&market_id).unwrap();
 		return market.get_market_price(outcome);
 	}
